API rest on core
