+++ src/main/java/br/unicamp/cst/core/entities/MemoryObject.java
@@ -12,10 +12,7 @@
 package br.unicamp.cst.core.entities;
 
 import java.io.Serializable;
-import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.Iterator;
-import java.util.List;
 import java.util.Set;
 
 /**
