+++ src/main/java/br/unicamp/cst/io/rest/HttpCodelet.java
@@ -2,5 +2,84 @@ package br.unicamp.cst.io.rest;
 
 import br.unicamp.cst.core.entities.Codelet;
 
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.OutputStream;
+import java.net.HttpURLConnection;
+import java.net.URL;
+
 public abstract class HttpCodelet extends Codelet {
+    /*public String URL;
+
+    public HttpCodelet(String URL){
+        super();
+        this.URL = URL;
+    }*/
+
+
+    public void sendPOST(String POST_URL, String POST_PARAMS) throws IOException {
+        URL obj = new URL(POST_URL);
+        HttpURLConnection con = (HttpURLConnection) obj.openConnection();
+        con.setRequestMethod("POST");
+        //con.setRequestProperty("User-Agent", USER_AGENT);
+
+        // For POST only - START
+        con.setDoOutput(true);
+        OutputStream os = con.getOutputStream();
+        os.write(POST_PARAMS.getBytes());
+        os.flush();
+        os.close();
+        // For POST only - END
+
+        int responseCode = con.getResponseCode();
+        System.out.println("POST Response Code :: " + responseCode);
+
+        if (responseCode == HttpURLConnection.HTTP_OK) { //success
+            BufferedReader in = new BufferedReader(new InputStreamReader(con.getInputStream()));
+            String inputLine;
+            StringBuffer response = new StringBuffer();
+
+            while ((inputLine = in.readLine()) != null) {
+                response.append(inputLine);
+            }
+            in.close();
+
+            // print result
+            System.out.println(response.toString());
+        } else {
+            System.out.println("POST request did not work.");
+        }
+    }
+
+    public String sendGET(String GET_URL) throws IOException {
+        String message = "";
+
+        URL obj = new URL(GET_URL);
+        HttpURLConnection con = (HttpURLConnection) obj.openConnection();
+        con.setRequestMethod("GET");
+        //con.setRequestProperty("User-Agent", USER_AGENT);
+        int responseCode=0;
+        try {
+            responseCode = con.getResponseCode();
+        } catch (java.net.ConnectException e) {
+            e.printStackTrace();
+        }
+        if (responseCode != 200)
+            System.out.println("GET Response Code :: " + responseCode);
+        if (responseCode == HttpURLConnection.HTTP_OK) { // success
+            BufferedReader in = new BufferedReader(new InputStreamReader(
+                    con.getInputStream()));
+            String inputLine;
+            StringBuffer response = new StringBuffer();
+            while ((inputLine = in.readLine()) != null) {
+                response.append(inputLine);
+            }
+            in.close();
+            message = response.toString();
+        } else {
+            System.out.println("GET request not worked");
+        }
+        return(message);
+    }
 }
