--- src/main/java/br/unicamp/cst/io/rest/MotorHttpCodelet.java
@@ -1,60 +1,0 @@
-package br.unicamp.cst.io.rest;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.OutputStream;
-import java.net.HttpURLConnection;
-import java.net.URL;
-
-public class MotorHttpCodelet extends HttpCodelet{
-    @Override
-    public void accessMemoryObjects() {
-
-    }
-
-    @Override
-    public void calculateActivation() {
-
-    }
-
-    @Override
-    public void proc() {
-
-    }
-
-    public void sendPOST(String POST_URL, String POST_PARAMS) throws IOException {
-        URL obj = new URL(POST_URL);
-        HttpURLConnection con = (HttpURLConnection) obj.openConnection();
-        con.setRequestMethod("POST");
-        //con.setRequestProperty("User-Agent", USER_AGENT);
-
-        // For POST only - START
-        con.setDoOutput(true);
-        OutputStream os = con.getOutputStream();
-        os.write(POST_PARAMS.getBytes());
-        os.flush();
-        os.close();
-        // For POST only - END
-
-        int responseCode = con.getResponseCode();
-        System.out.println("POST Response Code :: " + responseCode);
-
-        if (responseCode == HttpURLConnection.HTTP_OK) { //success
-            BufferedReader in = new BufferedReader(new InputStreamReader(con.getInputStream()));
-            String inputLine;
-            StringBuffer response = new StringBuffer();
-
-            while ((inputLine = in.readLine()) != null) {
-                response.append(inputLine);
-            }
-            in.close();
-
-            // print result
-            System.out.println(response.toString());
-        } else {
-            System.out.println("POST request did not work.");
-        }
-    }
-
-}
