+++ src/test/java/br/unicamp/cst/io/rest/TestRESTMemory.java
@@ -11,17 +11,19 @@ import java.util.Random;
 
 import static org.junit.Assert.assertEquals;
 
-public class TestRESTNew {
+public class TestRESTMemory {
     Mind m1;
     Mind m2;
 
     public Mind prepareMind(int portOut, int portIn, int partnerPortOut, int partnerPortIn, double outI, double toGetI) {
         //String URLOut = "http://localhost:"+portOut+"/";
         //String URLIn = "http://localhost:"+portIn+"/";
-        String partnerURLOut = "http://localhost:"+partnerPortOut+"/";
-        String partnerURLIn = "http://localhost:"+partnerPortIn+"/";
+        //String partnerURLOut = "http://localhost:"+partnerPortOut+"/";
+        String partnerURLOut = "http://127.0.0.1:"+partnerPortOut+"/";
+        //String partnerURLIn = "http://localhost:"+partnerPortIn+"/";
+        String partnerURLIn = "http://127.0.0.1:"+partnerPortIn+"/";
 
-        SensoryHttpCodelet restSensoryTestCodelet = new SensoryHttpCodelet(partnerURLOut) {
+        HttpCodelet restSensoryTestCodelet = new HttpCodelet() {
             //Memory
             @Override
             public void accessMemoryObjects() {
@@ -43,7 +45,7 @@ public class TestRESTNew {
             }
         };
 
-        MotorHttpCodelet restMotorTestCodelet = new MotorHttpCodelet() {
+        HttpCodelet restMotorTestCodelet = new HttpCodelet() {
             HashMap<String, String> params = new HashMap<>();
             final Random r = new Random();
             final Double I = outI; //(double) (5 + r.nextInt(500));
@@ -119,18 +121,22 @@ public class TestRESTNew {
         Mind m = new Mind();
 
 
-        RESTMemory m1 = new RESTMemory(portIn);
+        //RESTMemory m1 = new RESTMemory("127.0.0.1", portIn);
+        //m1.setI(1);
+        //m1.setName("M1");
+        //m.getRawMemory().addMemory(m1); //add manually to rawMemory
+        RESTMemory m1 = m.createRESTMemory("M1", "127.0.0.1", portIn);
         m1.setI(1);
-        m1.setName("M1");
-        m.getRawMemory().addMemory(m1); //add manually to rawMemory
 
         MemoryObject m2 = m.createMemoryObject("M2", 2.0);
         MemoryObject m3 = m.createMemoryObject("M3", null);
         MemoryObject m4 = m.createMemoryObject("M4", null);
-        RESTMemory m5 = new RESTMemory(portOut);
+        //RESTMemory m5 = new RESTMemory("127.0.0.1", portOut);
+        //m5.setI(toGetI);
+        //m5.setName("M5");
+        //m.getRawMemory().addMemory(m5);
+        RESTMemory m5 = m.createRESTMemory("M5", "127.0.0.1", portOut);
         m5.setI(toGetI);
-        m5.setName("M5");
-        m.getRawMemory().addMemory(m5);
 
         MemoryContainer m6 = m.createMemoryContainer("C1");
         MemoryContainer m7 = m.createMemoryContainer("C2");
@@ -176,7 +182,7 @@ public class TestRESTNew {
         int portOut1 = 5000 + r.nextInt(50000);
         int portOut2 = 5000 + r.nextInt(50000);
 
-        TestRESTNew tr = new TestRESTNew();
+        TestRESTMemory tr = new TestRESTMemory();
         double outI1 = (5 + r.nextInt(500));
         double toGetI1 = (5 + r.nextInt(500));
         double outI2 = (5 + r.nextInt(500));
